% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-templates.R
\name{template_basic}
\alias{template_basic}
\alias{template_osf_package}
\alias{template_shiny_package}
\title{Directory Templates}
\usage{
template_basic()

template_osf_package()

template_shiny_package()
}
\value{
a \code{tibble} object
}
\description{
Create a \code{tibble} object containing a file directory.
}
