% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_proj_dir_junctions.R
\name{mk_proj_dir_junction}
\alias{mk_proj_dir_junction}
\title{Make a Project Directory Junctions}
\usage{
mk_proj_dir_junction(template)
}
\arguments{
\item{template}{A \code{character} object specifying a file directory structure.
  The available templates are:
  \itemize{
    \item \code{'\link{template_basic}'}
    \item \code{'\link{template_osf_package}'}
    \item \code{'\link{template_shiny_package}'}
}}
}
\description{
\code{mk_proj_dir_junctions} creates a directory junction between pre-identified
  directories.
}
