% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_proj_dir_test.R
\name{mk_proj_dir_test}
\alias{mk_proj_dir_test}
\title{Make a project directory}
\usage{
mk_proj_dir_test()
}
\description{
\code{mk_proj_dir} creates a directory of folders for a generic project.
}
\details{
These folders will also contain \code{.gitkeep} or \code{.gitignore}
files (if necessary).
}
