% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_dir.R
\name{capture_dir}
\alias{capture_dir}
\title{Convert child directories into a tibble}
\usage{
capture_dir(root_path = "/home/rstudio/github/")
}
\arguments{
\item{root_path}{a character string of a parent directory}
}
\description{
\code{capture_dir} creates a \code{tibble} containing the path of each level
of child of a given parent directory.
}
