% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_shiny_pkg_dir.R
\name{mk_shiny_pkg_dir}
\alias{mk_shiny_pkg_dir}
\title{Make a Shiny Application Package directory}
\usage{
mk_shiny_pkg_dir()
}
\description{
\code{mk_shiny_pkg_dir} creates a directory of folders for a Shiny Application package.
}
\details{
These folders will also contain \code{/ui.R}, \code{/server.R}, and \code{/global.R}.
}
\note{
This file director comes from the a \href{https://community.rstudio.com/t/best-practices-shiny-development/1694/5?u=tiernan}{post} in the RStudio Community Forum
}
